feature(search): add elasticsearch and Store context with Search subcontext with product searching, filtering etc.

+++ src/Shared/Domain/ValueObjects/Money.php
@@ -6,5 +6,5 @@ namespace Src\Shared\Domain\ValueObjects;
 
 final readonly class Money
 {
-    public function __construct(public float $amount) {}
+    public function __construct(public int $amount) {}
 }
